All funcitons in SVONVolume implemented. Build warning fixed

+++ SVON/Public/SVONData.h
@@ -22,7 +22,7 @@ namespace SVON
 		{
 			int result = 0;
 			result += leafNodes.size() * sizeof(SVONLeafNode);
-			for (int i = 0; i < layers.size(); ++i)
+			for (size_t i = 0; i < layers.size(); ++i)
 			{
 				result += layers[i].size() * sizeof(SVONNode);
 			}
